+++ src/app/Playground/HåvardSamples/HåvardPage.xaml.cs
@@ -2,16 +2,8 @@ namespace Playground.HåvardSamples;
 
 public partial class HåvardPage
 {
-    private int m_numberOfTimesTapped;
-
     public HåvardPage()
     {
         InitializeComponent();
-    }
-
-    private void NavigationListItem_OnTapped(object sender, EventArgs e)
-    {
-        label.Text = m_numberOfTimesTapped.ToString();
-        m_numberOfTimesTapped++;
     }
 }